+++ src/inc/macho.hpp
@@ -33,11 +33,11 @@ class Macho : public ExecutableFormat
 
         std::vector<std::shared_ptr<Section>> get_executables_section(std::ifstream & file);
 
-        unsigned long long raw_offset_to_va(const unsigned long long absolute_raw_offset, const unsigned long long absolute_raw_offset_section) const;
+        uint64_t raw_offset_to_va(const uint64_t absolute_raw_offset, const uint64_t absolute_raw_offset_section) const;
 
         void display_information(const VerbosityLevel lvl) const;
 
-        unsigned long long get_image_base_address(void);
+        uint64_t get_image_base_address(void);
 
     private:
