+++ src/inc/pe.hpp
@@ -36,7 +36,7 @@ class PE : public ExecutableFormat
 
         std::vector<std::shared_ptr<Section>> get_executables_section(std::ifstream & file);
 
-        unsigned long long get_image_base_address(void);
+        uint64_t get_image_base_address(void);
 
     private:
